// api-tests/interfaces/booking.interface.ts
export interface BookingDates {
    checkin: string;
    checkout: string;
}

export interface Booking {
    firstname: string;
    lastname: string;
    totalprice: number;
    depositpaid: boolean;
    bookingdates: BookingDates;
    additionalneeds: string;
}

export interface BookingResponse {
    bookingid: number;
    booking: Booking;
}